backend/requirements.txt
fastapi
uvicorn
pypdf
chromadb
python-dotenv
sentence-transformers
python-multipart
pdf2image
pytesseract
pytest
rank-bm25
groq
